nlp_chatbot/requirements.txt
nltk==3.9.1
spacy==3.8.4
flask==3.1.0
python-dotenv==1.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
